Restructured features into smaller chunks.
